chore(platform): share channel test org/home helpers
Deduplicate createDefaultTestOrgs, createMultiTestOrgs (Acme/Beta), createTestOrgStore, and withTempHome across telegram/whatsapp/discord test helpers.
